+++ CRM/Campaign/BAO/Petition.php
@@ -588,7 +588,7 @@ AND         tag_id = ( SELECT id FROM civicrm_tag WHERE name = %2 )";
       case CRM_Campaign_Form_Petition_Signature::EMAIL_CONFIRM:
         // create mailing event subscription record for this contact
         // this will allow using a hash key to confirm email address by sending a url link
-        $se = CRM_Mailing_Event_BAO_Subscribe::subscribe($group_id,
+        $se = CRM_Mailing_Event_BAO_MailingEventSubscribe::subscribe($group_id,
           $params['email-Primary'],
           $params['contactId'],
           'profile'
