+++ CRM/Mailing/BAO/MailingJob.php
@@ -368,7 +368,7 @@ class CRM_Mailing_BAO_MailingJob extends CRM_Mailing_DAO_MailingJob {
    * @param int $offset
    */
   public function split_job($offset = 200) {
-    $recipient_count = CRM_Mailing_BAO_Recipients::mailingSize($this->mailing_id);
+    $recipient_count = CRM_Mailing_BAO_MailingRecipients::mailingSize($this->mailing_id);
 
     $jobTable = CRM_Mailing_DAO_MailingJob::getTableName();
 
@@ -421,7 +421,7 @@ VALUES (%1, %2, %3, %4, %5, %6, %7)
     else {
       // We are still getting all the recipients from the parent job
       // so we don't mess with the include/exclude logic.
-      $recipients = CRM_Mailing_BAO_Recipients::mailingQuery($this->mailing_id, $this->job_offset, $this->job_limit);
+      $recipients = CRM_Mailing_BAO_MailingRecipients::mailingQuery($this->mailing_id, $this->job_offset, $this->job_limit);
 
       // FIXME: this is not very smart, we should move this to one DB call
       // INSERT INTO ... SELECT FROM ..
@@ -454,14 +454,14 @@ VALUES (%1, %2, %3, %4, %5, %6, %7)
         $count++;
         // dev/core#1768 Mail sync interval is now configurable.
         if ($count % $mail_sync_interval == 0) {
-          CRM_Mailing_Event_BAO_Queue::bulkCreate($params, $now);
+          CRM_Mailing_Event_BAO_MailingEventQueue::bulkCreate($params, $now);
           $count = 0;
           $params = [];
         }
       }
 
       if (!empty($params)) {
-        CRM_Mailing_Event_BAO_Queue::bulkCreate($params, $now);
+        CRM_Mailing_Event_BAO_MailingEventQueue::bulkCreate($params, $now);
       }
     }
   }
@@ -679,7 +679,7 @@ VALUES (%1, %2, %3, %4, %5, %6, %7)
         $params = array_merge($params,
           CRM_Mailing_BAO_BouncePattern::match($result->getMessage())
         );
-        CRM_Mailing_Event_BAO_Bounce::create($params);
+        CRM_Mailing_Event_BAO_MailingEventBounce::create($params);
       }
       elseif (is_a($result, 'PEAR_Error') && $mailing->sms_provider_id) {
         // Handle SMS errors: CRM-15426
@@ -936,7 +936,7 @@ AND    status IN ( 'Scheduled', 'Running', 'Paused' )
     static $writeActivity = NULL;
 
     if (!empty($deliveredParams)) {
-      CRM_Mailing_Event_BAO_Delivered::bulkCreate($deliveredParams);
+      CRM_Mailing_Event_BAO_MailingEventDelivered::bulkCreate($deliveredParams);
       $deliveredParams = [];
     }
 
@@ -1037,17 +1037,17 @@ AND    record_type_id = $targetRecordID
    * @param string $medium
    *   Ex: 'email' or 'sms'.
    *
-   * @return \CRM_Mailing_Event_BAO_Queue
+   * @return \CRM_Mailing_Event_BAO_MailingEventQueue
    *   A query object whose rows provide ('id', 'contact_id', 'hash') and ('email' or 'phone').
    */
   public static function findPendingTasks($jobId, $medium) {
-    $eq = new CRM_Mailing_Event_BAO_Queue();
-    $queueTable = CRM_Mailing_Event_BAO_Queue::getTableName();
+    $eq = new CRM_Mailing_Event_BAO_MailingEventQueue();
+    $queueTable = CRM_Mailing_Event_BAO_MailingEventQueue::getTableName();
     $emailTable = CRM_Core_BAO_Email::getTableName();
     $phoneTable = CRM_Core_BAO_Phone::getTableName();
     $contactTable = CRM_Contact_BAO_Contact::getTableName();
-    $deliveredTable = CRM_Mailing_Event_BAO_Delivered::getTableName();
-    $bounceTable = CRM_Mailing_Event_BAO_Bounce::getTableName();
+    $deliveredTable = CRM_Mailing_Event_BAO_MailingEventDelivered::getTableName();
+    $bounceTable = CRM_Mailing_Event_BAO_MailingEventBounce::getTableName();
 
     $query = "  SELECT      $queueTable.id,
                                 $emailTable.email as email,
