+++ CRM/Mailing/Form/Subscribe.php
@@ -143,7 +143,7 @@ ORDER BY title";
       }
     }
 
-    CRM_Mailing_Event_BAO_Subscribe::commonSubscribe($groups, $params);
+    CRM_Mailing_Event_BAO_MailingEventSubscribe::commonSubscribe($groups, $params);
   }
 
 }
