+++ tests/phpunit/CRM/Mailing/BAO/QueryTest.php
@@ -69,8 +69,8 @@ class CRM_Mailing_BAO_QueryTest extends CiviUnitTestCase {
     $this->loadXMLDataSet(dirname(__FILE__) . '/queryDataset.xml');
     // ensure that total unique opened mail count is same while
     //   fetching rows and row count for mailing_id = 14
-    $totalOpenedMailCount = CRM_Mailing_Event_BAO_Opened::getTotalCount(14, NULL, TRUE);
-    $totalOpenedMail = CRM_Mailing_Event_BAO_Opened::getRows(14, NULL, TRUE);
+    $totalOpenedMailCount = CRM_Mailing_Event_BAO_MailingEventOpened::getTotalCount(14, NULL, TRUE);
+    $totalOpenedMail = CRM_Mailing_Event_BAO_MailingEventOpened::getRows(14, NULL, TRUE);
 
     $this->assertEquals(4, $totalOpenedMailCount);
     $this->assertEquals(4, count($totalOpenedMail));
@@ -85,9 +85,9 @@ class CRM_Mailing_BAO_QueryTest extends CiviUnitTestCase {
     // ensure that total unique clicked mail count is same while
     //   fetching rows and row count for mailing_id = 14 and
     //   trackable_url_id 12
-    $totalDistinctTrackableUrlCount = CRM_Mailing_Event_BAO_TrackableURLOpen::getTotalCount(14, NULL, TRUE, 13);
-    $totalTrackableUrlCount = CRM_Mailing_Event_BAO_TrackableURLOpen::getTotalCount(14, NULL, FALSE, 13);
-    $totalTrackableUrlMail = CRM_Mailing_Event_BAO_TrackableURLOpen::getRows(14, NULL, TRUE, 13);
+    $totalDistinctTrackableUrlCount = CRM_Mailing_Event_BAO_MailingEventClickThrough::getTotalCount(14, NULL, TRUE, 13);
+    $totalTrackableUrlCount = CRM_Mailing_Event_BAO_MailingEventClickThrough::getTotalCount(14, NULL, FALSE, 13);
+    $totalTrackableUrlMail = CRM_Mailing_Event_BAO_MailingEventClickThrough::getRows(14, NULL, TRUE, 13);
 
     $this->assertEquals(3, $totalDistinctTrackableUrlCount, "Accurately display distinct count of unique trackable URLs");
     $this->assertEquals(4, $totalTrackableUrlCount, "Accurately display count of unique trackable URLs");
